+++ DualityEditor/DualityEditor.csproj
@@ -114,8 +114,9 @@
     <Compile Include="Backend\IEditorGraphicsBackend.cs" />
     <Compile Include="Controls\TreeModels\TypeHierarchy\ModelItems.cs" />
     <Compile Include="Controls\TreeModels\TypeHierarchy\TypeBrowserTreeModel.cs" />
-    <Compile Include="Extensibility\AssetImport\AssetReImportOperation.cs" />
     <Compile Include="Extensibility\AssetImport\AssetImportOperation.cs" />
+    <Compile Include="Extensibility\AssetImport\AssetReImportOperation.cs" />
+    <Compile Include="Extensibility\AssetImport\AssetFirstImportOperation.cs" />
     <Compile Include="Extensibility\AssetImport\AssetImportInput.cs" />
     <Compile Include="Extensibility\AssetImport\ExtMethodsIAssetImportEnvironment.cs" />
     <Compile Include="Extensibility\DataConversion\ConversionData.cs" />
