+++ DualityEditorPlugins/EditorBase/Importers/PixmapAssetImporter.cs
@@ -82,4 +82,77 @@ namespace Duality.Editor.Plugins.Base
 			return pixelData;
 		}
 	}
+
+	public class PixmapAssetImporter2 : IAssetImporter
+	{
+		public static readonly string SourceFileExtPrimary = ".png";
+		private static readonly string[] SourceFileExts = new[] { SourceFileExtPrimary, ".bmp", ".jpg", ".jpeg", ".tif", ".tiff" };
+		
+		public string Id
+		{
+			get { return "BasicPixmapAssetImporter2"; }
+		}
+		public string Name
+		{
+			get { return "Pixmap Importer 2"; }
+		}
+		public int Priority
+		{
+			get { return -1; }
+		}
+
+		public void PrepareImport(IAssetImportEnvironment env)
+		{
+			// Ask to handle all input that matches the conditions in AcceptsInput
+			foreach (AssetImportInput input in env.HandleAllInput(this.AcceptsInput))
+			{
+				// For all handled input items, specify which Resource the importer intends to create / modify
+				env.AddOutput<Pixmap>(input.AssetName);
+			}
+		}
+		public void Import(IAssetImportEnvironment env) { }
+		
+		private bool AcceptsInput(AssetImportInput input)
+		{
+			string inputFileExt = Path.GetExtension(input.Path);
+			bool matchingFileExt = SourceFileExts.Any(acceptedExt => string.Equals(inputFileExt, acceptedExt, StringComparison.InvariantCultureIgnoreCase));
+			return matchingFileExt;
+		}
+	}
+	public class PixmapAssetImporter3 : IAssetImporter
+	{
+		public static readonly string SourceFileExtPrimary = ".png";
+		private static readonly string[] SourceFileExts = new[] { SourceFileExtPrimary, ".bmp", ".jpg", ".jpeg", ".tif", ".tiff" };
+		
+		public string Id
+		{
+			get { return "BasicPixmapAssetImporter3"; }
+		}
+		public string Name
+		{
+			get { return "Pixmap Importer 3"; }
+		}
+		public int Priority
+		{
+			get { return -2; }
+		}
+
+		public void PrepareImport(IAssetImportEnvironment env)
+		{
+			// Ask to handle all input that matches the conditions in AcceptsInput
+			foreach (AssetImportInput input in env.HandleAllInput(this.AcceptsInput))
+			{
+				// For all handled input items, specify which Resource the importer intends to create / modify
+				env.AddOutput<Pixmap>(input.AssetName);
+			}
+		}
+		public void Import(IAssetImportEnvironment env) { }
+		
+		private bool AcceptsInput(AssetImportInput input)
+		{
+			string inputFileExt = Path.GetExtension(input.Path);
+			bool matchingFileExt = SourceFileExts.Any(acceptedExt => string.Equals(inputFileExt, acceptedExt, StringComparison.InvariantCultureIgnoreCase));
+			return matchingFileExt;
+		}
+	}
 }
